refactor: extract shared UI toolkit into src/ui.rs
Move the reusable widget builders + helpers (btn_primary/btn_secondary,
tab_pill, menu_icon, pack_ext/pack_size, lerp_rgb, scrollbar_thumb) out of
render.rs into src/ui.rs, re-exported at the crate root so every feature
module reaches them via `use super::*`. Prereq for splitting render_*
methods into per-feature files. scrollbar_thumb's panic-safety test moves
with it. Zero behaviour change.
